requirements.txt
Flask==3.0.2
twilio==6.50.1
openai
python-dotenv